Fix | Moving RequestBoundaryMethodsTest.java to connection package
