+++ src/test/java/com/microsoft/sqlserver/jdbc/connection/RequestBoundaryMethodsTest.java
@@ -1,4 +1,4 @@
-package com.microsoft.sqlserver.jdbc.requestboundary;
+package com.microsoft.sqlserver.jdbc.connection;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
